x64: pooling: fix residual dim_t->int narrowing in apply_postops
compute_vector_range takes int vmm indices (semantic range 0-31);
start_idx/end_idx are dim_t from surrounding arithmetic and need an
explicit boundary cast, same pattern used elsewhere for vmm indices.

+++ src/cpu/x64/jit_uni_pool_kernel.cpp
@@ -800,7 +800,8 @@ void jit_uni_pool_kernel_t<isa>::apply_postops(dim_t ur_bc, dim_t ur_w,
             }
         }
     }
-    postops_injector_->compute_vector_range(start_idx, end_idx, rhs_arg_params);
+    postops_injector_->compute_vector_range(static_cast<int>(start_idx),
+            static_cast<int>(end_idx), rhs_arg_params);
 }
 
 template <cpu_isa_t isa>
